pyproject.toml
[tool.poetry]
name = "projeto-fiap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
setuptools = "^72.2.0"
pyyaml = "^6.0.2"
uritemplate = "^4.1.1"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
